Move 'FeedViewController' to production
